backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend service for the application"
readme = "README.md"
authors = [
    { name = "M", email = "lmiguelvargasf@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "advanced-alchemy>=1.8.2",
    "argon2-cffi",
    "asyncpg",
    "itsdangerous",
    "litestar[standard]>=2.19.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "sqladmin>=0.22.0",
    "strawberry-graphql>=0.287.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyrefly>=0.46.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
    "strawberry-graphql[debug-server]>=0.287.3",
]

[tool.ruff]
line-length = 88
target-version = "py314"
[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "RUF"]
ignore = ["E501"]
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "--cov=src"

[tool.coverage.run]
omit = [
]

[tool.pyrefly]
python-version = "3.14"
project-includes = ["src"]
search-path = ["src"]
site-package-path = [".venv/lib/python3.14/site-packages"]
